CHROMIUM: mmc: sdhci: allow sdio interrupts while sdhci rt suspended
(In the spirit of upstream commit
be138554a7923658ded799b0e8794d9c1d08a6e5; However, with some
prerequisites to make it function without the rest of the
re-architecting that went along with it.  Note the sdhci-esdhc-imx
component is also omitted as RTPM is not implemented for this driver
in 3.10)

Allow SDIO interrupts to be received while the SDHCI host is runtime
suspended.  Host will need to leave any clocks on as required to read
SDHCI registers, particularly the interrupt status register.

BUG=chrome-os-partner:32023
TEST=Bluetooth keyboard keypresses after idling (Wifi disabled)

Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>

Change-Id: I7be011ce1a13fae82664dc3e1382f31081821326
Reviewed-on: https://chromium-review.googlesource.com/233320
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Christopher Freeman <cfreeman@nvidia.com>
Tested-by: Christopher Freeman <cfreeman@nvidia.com>